get int64 values

+++ Convert.go
@@ -2,9 +2,9 @@ package bunit
 
 import "github.com/cymertek/go-big"
 
-func (b Bytes) Int() int64 {
+func (b Bytes) Int64() int64 {
 	return (&big.Int{}).SetBytes(b).Int64()
 }
-func (b Bits) Int() int64 {
+func (b Bits) Int64() int64 {
 	return (&big.Int{}).SetBytes(b).Int64()
 }
